app/requirements.txt
flask==3.0.0
mysql-connector-python==8.3.0
prometheus-flask-exporter==0.23.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-flask==0.43b0
pytest==7.4.0
